// src/constants/enums/RequestMethod.ts
/**
 * Enum representing HTTP request methods.
 * @enum {string}
 * @ordinal {string} POST - The POST request method.
 * @ordinal {string} GET - The GET request method.
 * @ordinal {string} HEAD - The HEAD request method.
 * @ordinal {string} PUT - The PUT request method.
 * @ordinal {string} PATCH - The PATCH request method.
 * @ordinal {string} DELETE - The DELETE request method.
 */
export enum RequestMethod {
    POST = "POST",
    GET = "GET",
    HEAD = "HEAD",
    PUT = "PUT",
    PATCH = "PATCH",
    DELETE = "DELETE",
};